+++ src/d_main.c
@@ -148,8 +148,9 @@ boolean respawnparm;    // working -respawn
 boolean fastparm;       // working -fast
 
 // [Nugget]
-boolean doubleammoparm;
-boolean halfdamageparm;
+boolean coopspawnsparm = false;
+boolean doubleammoparm = false;
+boolean halfdamageparm = false;
 
 boolean singletics = false; // debug flag to cancel adaptiveness
 
@@ -2040,7 +2041,17 @@ static boolean CheckHaveSSG (void)
   return true;
 }
 
-// [Nugget]
+// [Nugget] /-----------------------------------------------------------------
+
+void D_ValidateStartSkill(void)
+{
+  if (startskill == sk_custom
+      && (demorecording || demoplayback || netgame || strictmode))
+  {
+    startskill = sk_hard;
+  }
+}
+
 void D_NuggetUpdateCasual(void)
 {
   static int old_casual = -1;
@@ -2058,6 +2069,8 @@ void D_NuggetUpdateCasual(void)
     MN_SetupResetMenu();
   }
 }
+
+// [Nugget] -----------------------------------------------------------------/
 
 //
 // D_DoomMain
@@ -2395,28 +2408,30 @@ void D_DoomMain(void)
   startmap = 1;
   autostart = false;
 
+  // [Nugget] Take custom skill into account below
+
   //!
   // @category game
   // @arg <skill>
   // @vanilla
   // @help
   //
-  // Set the game skill, 1-5 (1: easiest, 5: hardest). A skill of 0 disables all
-  // monsters only in -complevel vanilla.
+  // Set the game skill, 1-6 (1: easiest, 5: hardest, 6: custom).
+  // A skill of 0 disables all monsters only in -complevel vanilla.
   //
 
   if ((p = M_CheckParm ("-skill")) && p < myargc-1)
    {
      startskill = M_ParmArgToInt(p);
      startskill--;
-     if (startskill >= sk_none && startskill <= sk_nightmare)
+     if (startskill >= sk_none && startskill <= sk_custom)
       {
         autostart = true;
       }
      else
       {
-        I_Error("Invalid parameter '%s' for -skill, valid values are 1-5 "
-                "(1: easiest, 5: hardest).\n"
+        I_Error("Invalid parameter '%s' for -skill, valid values are 1-6\n"
+                "(1: easiest, 5: hardest, 6: custom).\n"
                 "In complevel Vanilla, '-skill 0' disables all monsters.", myargv[p+1]);
       }
    }
@@ -2802,6 +2817,8 @@ void D_DoomMain(void)
   // Initial netgame startup. Connect to server etc.
   D_ConnectNetGame();
 
+  D_ValidateStartSkill(); // [Nugget]
+
   I_Printf(VB_INFO, "D_CheckNetGame: Checking network game status.");
   D_CheckNetGame();
 
@@ -2849,7 +2866,7 @@ void D_DoomMain(void)
 
   if (M_ParmExists("-coop_spawns"))
     {
-      coop_spawns = true;
+      coopspawnsparm = true;
     }
 
   //!
@@ -2971,6 +2988,8 @@ void D_DoomMain(void)
   {
     I_SetFastdemoTimer(true);
   }
+
+  D_ValidateStartSkill(); // [Nugget]
 
   // [FG] init graphics (video.widedelta) before HUD widgets
   I_InitGraphics();