+++ src/doomdef.h
@@ -120,7 +120,9 @@ typedef enum {
   sk_easy,
   sk_medium,
   sk_hard,
-  sk_nightmare
+  sk_nightmare,
+
+  sk_custom // [Nugget]
 } skill_t;
 
 //
@@ -196,6 +198,9 @@ typedef enum {
   ss_enem,
   ss_gen,       // killough 10/98
   ss_comp,      // killough 10/98
+
+  ss_skill, // [Nugget] Custom Skill menu
+
   ss_max
 } ss_types;