+++ src/m_cheat.c
@@ -1272,15 +1272,12 @@ static void cheat_skill(char *buf)
 {
   int skill = buf[0] - '0';
 
-  if (skill >= 1 && skill <= 5)
+  if (skill >= 1 && skill <= 6) // [Nugget] Custom Skill
   {
     gameskill = skill - 1;
     displaymsg("Next Level Skill: %s", default_skill_strings[gameskill + 1]);
 
-    G_SetFastParms(fastparm || gameskill == sk_nightmare);
-    respawnmonsters = gameskill == sk_nightmare || respawnparm;
-
-    G_SetBabyModeParms(gameskill); // [Nugget]
+    G_SetSkillParms(gameskill); // [Nugget]
   }
 }