+++ src/p_mobj.c
@@ -964,7 +964,7 @@ mobj_t *P_SpawnMobj(fixed_t x, fixed_t y, fixed_t z, mobjtype_t type)
 
   mobj->health = info->spawnhealth;
 
-  if (gameskill != sk_nightmare)
+  if (!aggressive) // [Nugget] Custom Skill: use `aggressive`
     mobj->reactiontime = info->reactiontime;
 
   mobj->lastlook = P_Random (pr_lastlook) % MAXPLAYERS;
@@ -1361,10 +1361,11 @@ void P_SpawnMapThing (mapthing_t* mthing)
     return;
 
   // killough 11/98: simplify
+  // [Nugget] Custom Skill: use `thingspawns`
   if ((gameskill == sk_none && demo_compatibility) ||
-      (gameskill == sk_baby || gameskill == sk_easy ?
+      (thingspawns == THINGSPAWNS_EASY ?
       !(mthing->options & MTF_EASY) :
-      gameskill == sk_hard || gameskill == sk_nightmare ?
+      thingspawns == THINGSPAWNS_HARD ?
       !(mthing->options & MTF_HARD) : !(mthing->options & MTF_NORMAL)))
     return;